requirements.txt
OpenCV-Python
MediaPipe
imutils
cv2-enumerate-cameras